\name{CoveragePlot}
\alias{CoveragePlot}
\title{Coverage Plot for Gene and Transcript Regions using BAM files}
\description{
  This function generates a gene region track and corresponding read coverage
  plots from BAM files using the \code{Gviz} package. The function takes
  gene and transcript annotations from a GTF file and BAM files as input,
  and produces a visual plot of read coverage and gene/transcript regions.
}
\usage{
CoveragePlot(bamfiles, gtf, gene_name, transcript_names, extend.range = 0,
             genome = "mm10", zoom = FALSE, zoom_focus = 0.1)
}
\arguments{
  \item{bamfiles}{
    A named vector of BAM file paths. The names will be used as labels in the
    plot.
  }
  \item{gtf}{
    A \code{GRanges} object containing the GTF annotation data.
  }
  \item{gene_name}{
    The name of the gene of interest, which must be present in the GTF data.
  }
  \item{transcript_names}{
    A vector of transcript names to be plotted. These must match transcripts
    present in the GTF data.
  }
  \item{extend.range}{
    A numeric value specifying the extension of the plot region beyond the
    gene's boundaries. The default is 0 (no extension).
  }
  \item{genome}{
    The genome build identifier to be used (e.g., "mm10", "hg19").
    The default is "mm10".
  }
  \item{zoom}{
    Logical. Should the plot zoom into a specific region? The default is
    \code{FALSE}.
  }
  \item{zoom_focus}{
    A numeric value representing the percentage of the gene region to focus
    on when zooming. The default is 0.1 (i.e., 10\% of the region).
  }
}
\value{
  A plot generated by \code{Gviz} that displays gene and transcript
  annotations along with read coverage tracks from the provided BAM files.
}
\details{
  This function visualizes the gene and transcript structures using the
  \code{GeneRegionTrack} function from the \code{Gviz} package, and adds read
  coverage tracks using \code{AlignmentsTrack}. The plot shows the genomic
  features of the specified gene and transcripts, overlaid with read coverage
  information from BAM files.
  If zooming is enabled, the plot will focus on a smaller region of the gene.
}
\seealso{
  \code{\link[Gviz]{GeneRegionTrack}},
  \code{\link[Gviz]{AlignmentsTrack}},
  \code{\link[GenomicRanges]{GRanges}}
}
\author{
  Franz AKE (fake@idibell.cat)
}
\keyword{visualization}
\keyword{genomics}
